Update Uniti-Tim-Unerman.md
Just a couple of edits here and there, making the format of paragraphs with zero spaces between the heading and the question, so it looks more professional. It's better to keep the same formatting for each question with regards to where Tim's response is on the page; example - underneath the question, next to the question, a paragraph below the question.
I've left the spaces in that you have put in between some of his answers, please can you tell me, is that to make particular sentences stand out? Perhaps we could put some pull quotes in instead, this is a quote in a heading format that separates interview questions to make it easier to read and draws in the readers' attention. 
Also just putting capitals at the beginning of sentences, some of them have seemed to have transformed into lower case! Just i's and t's.